+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,13 @@
 # seroval-plugins
 
+## 1.4.2
+
+## 1.4.1
+
+### Patch Changes
+
+- 3995cc1: security enhancements
+
 ## 1.4.0
 
 ### Minor Changes