+++ packages/seroval/CHANGELOG.md
@@ -1,5 +1,17 @@
 # seroval
 
+## 1.4.2
+
+### Patch Changes
+
+- fix byte lengths
+
+## 1.4.1
+
+### Patch Changes
+
+- 3995cc1: security enhancements
+
 ## 1.4.0
 
 ### Minor Changes